# backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: Cash
  description: Cash app
  version: 1.0.0
servers:
- url: /
...
paths:
  /user:
    post:
      tags:
      - User
      summary: Create user
      description: This can only be done once at the registration page
      operationId: createUser
      requestBody:
          content:
              application/json:
                  schema:
                      $ref: "#/components/schemas/User"
      responses:
        "200":
          description: "successful operation, user is logged in"
        "400":
          description: Invalid email supplied
        "413":
          description: Invalid password supplied
      x-swagger-router-controller: User
    put:
      tags:
      - User
      summary: Update user
      description: This can only be done by the logged in user.
      operationId: updateUser
      security:
        - bearerAuth: []
      requestBody:
          content:
              application/json:
                  schema:
                      $ref: "#/components/schemas/User"
      responses:
        "400":
          description: Invalid user supplied
        "404":
          description: User not found
      x-swagger-router-controller: User
    delete:
      tags:
      - User
      summary: Delete user
      description: This can only be done by the logged in user
      operationId: deleteUser
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User has been deleted
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    User:
      type: object
      properties:
        userId:
          type: integer
        email:
          type: string
        password:
          type: string
          format: password
        name:
          type: string
      required:
        - email
        - password
        - name
        
      xml:
        name: User
...